Refactors constants and enums for better modularity
Moves enums to a separate types file to enhance organization
and reusability across modules. Updates imports and improves
typing for environment and output format defaults. Ensures
consistent usage of centralized constants and cleanup of
related code.

Relates to gitflow-feature-stash: export-const-types
