+++ src/index.ts
@@ -1,5 +1,6 @@
 export { logger } from './logger'; // No requiere cambio, solo si se exporta desde index.ts como @smdv/logger
 export { XmlProcessor } from './xml';
 export { requestLogger } from './middleware';
-export { LogLevel, LoggerConfig, Transport, HttpStatusCode, ApplicationErrorCode, ErrorContext } from './types';
+export { LogLevel, OutputFormat, Environment, SupportedLang, LoggerConfig, Transport, HttpStatusCode, ApplicationErrorCode, ErrorContext } from './types';
+export { ENV_KEYS, DEFAULTS } from './constants';
 export { createCustomLogger } from './factory';
